Isolated the pandas-dependent code in _dataframe.py.
